added complexity tests

+++ LogoIntersectionFinder/Helpers/Vector.cs
@@ -2,14 +2,14 @@
 {
     public class Vector
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public long X { get; set; }
+        public long Y { get; set; }
         public Vector(int x, int y)
         {
             X = x;
             Y = y;
         }
-        public int CrossProduct(Vector v)
+        public long CrossProduct(Vector v)
         {
             return X * v.Y - v.X * Y;
         }
